+++ LeetCode/441_排列硬币.cpp
@@ -6,6 +6,6 @@ class Solution {
 public:
     int arrangeCoins(int n) {
         long long delta = static_cast<long long>(sqrt(1L + 8L * n));
-        return (delta - 1) / 2;
+        return static_cast<int>((delta - 1) / 2);
     }
 };
